Move assets out of top-level to prevent conflicts
I ran into a problem when including wysiwyg-rails into our project.
After adding the gem to our Gemfile and attempting to load a page, we
were getting an error about an undefined sass variable.

The problem is that wysiwyg-rails adds themes/*.css to the asset
precompilation array and our project also has a themes directory. But
unlike froala's themes directory ours has partials that can't be
compiled by themselves, so we'd get a sass error when trying to view any
page.