Remove assets precompile config initializer
This is unnecessary as an asset gem. As shown in the docs for wysiwyg-rails, these files are required in manifest files that are part of a Rails applications' precompile config. There is no need to also precompile these files too which causes a very large amount of files being generated when running `rails assets:precompile` (an extra file for each due to GZ).